% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bar_plot_contingency_tables}
\alias{bar_plot_contingency_tables}
\title{bar plot for contingency tables}
\usage{
bar_plot_contingency_tables(
  tables,
  rejected = FALSE,
  scale_factor = 1,
  add_legend = FALSE,
  legend_ncol = 2,
  legend_cex = 1,
  ...
)
}
\arguments{
\item{tables}{list contingency tables.}

\item{rejected}{TRUE if number of rejected hypothesis is plotted in y axis.}

\item{scale_factor}{a numeric factor that is multiplied to data values.}

\item{add_legend}{If not FALSE, legend is added in the plot}

\item{legend_ncol}{ncol for the legend}

\item{legend_cex}{cex for the legend}

\item{...}{Additional arguments to barplot()}
}
\description{
bar plot for contingency tables
}
\examples{
  report <- simulate_fragment_ion_report(default_params)
  resu <- run_ttests(report, boot_denom_eps = 0.3)
  tables <- compute_contingency_tables(resu, alpha = 0.05)
  bar_plot_contingency_tables(
    tables[1], xlab = "Comparison", ylab = "1 - Type I error rate",
    cex.lab = 1.5, add_legend = TRUE
  )
}
