% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{ca_normalize_values_df}
\alias{ca_normalize_values_df}
\title{Normalize values to standard normal distribution by category}
\usage{
ca_normalize_values_df(
  values_df,
  value_column = "precursor_quantity",
  category_column = "condition",
  reconstruct = TRUE,
  use_logvalues = TRUE,
  normalized_value_column = "normalized_precursor_quantity",
  normalization_factor_column = "normalization_factor"
)
}
\arguments{
\item{values_df}{data frame with value_column and category_column}

\item{value_column}{name of the value column}

\item{category_column}{name of the category column}

\item{reconstruct}{If TRUE, the normalized_value_column is transformed to
take the robust mean and robust standard deviation of the value_column.
If FALSE, the mean and standard deviation are set to 0 and 1.}

\item{use_logvalues}{If TRUE, the value column is log10-transformed,
normalized, and then transformed back to the original exponential scale.}

\item{normalized_value_column}{name of the normalized value column}

\item{normalization_factor_column}{name of the normalization factor column}
}
\value{
data frame with normalized values set in the new column
}
\description{
Normalize values to standard normal distribution by category
}
