% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_contingency_tables}
\alias{compute_contingency_tables}
\title{Compute contingency tables}
\usage{
compute_contingency_tables(
  test_results,
  alpha = 0.05,
  q_value_column = "p.value"
)
}
\arguments{
\item{test_results}{Results of run_ttest function, which is a named
list of analysis results of each method.}

\item{alpha}{Significance level}

\item{q_value_column}{Name of the column having significance levels: A
smaller value implies more significance.}
}
\value{
List of contingency tables for every comparisons. Each element of
the list is a table, whose columns are the t-test methods and the rows are
FALSE/TRUE, where FALSE represents that null hypothesis is not rejected,
.  and TRUE represents rejected.
}
\description{
Compute contingency tables based on the results of running t-test methods.
}
