% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_cov_unequal_replicates}
\alias{compute_cov_unequal_replicates}
\title{Compute cov_unequal_replicates}
\usage{
compute_cov_unequal_replicates(report_df)
}
\arguments{
\item{report_df}{a fragment ion report with multiple conditions and the
fragment_peak_area column}
}
\value{
the report data frame with a new column "cov_unequal_replicates"
}
\description{
Compute cov_unequal_replicates by each protein, and append it into a new
column. The fragment_ion_peak area is first summarized to
log10_peptide_quantity in precursor level. It is then used for estimating
a mixture of normal and expbeta distributions by each protein.
}
