% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_t_test.R
\name{compute_indep_on_stdreport}
\alias{compute_indep_on_stdreport}
\title{Run independent samples t-test comparing two conditions on a standard report}
\usage{
compute_indep_on_stdreport(report)
}
\arguments{
\item{report}{fragment ion report with the columns experiment, condition,
replicate, protein_id, precursor_id, and fragment_peak_area.
The condition column is supposed to consist of two conditions.}
}
\value{
a data frame of the test results. Columns are independent_t_test
results, and rows are precursors.
}
\description{
Run independent samples t-test comparing two conditions on a standard report
}
