% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{compute_lfdr_result}
\alias{compute_lfdr_result}
\title{Append local false discovery (lfdr) score to the statistical testing results.}
\usage{
compute_lfdr_result(
  ttest_result,
  known_target_df = NULL,
  is_target_column = "is_target"
)
}
\arguments{
\item{ttest_result}{A list of lists of statistical test result tables,
which are obtained by run_ttest().}

\item{known_target_df}{An optional data frame for known target data frame.
The 'UniprotID' column at the first column is the list of uniprot IDs.}

\item{is_target_column}{Name of the boolean column for indicating on-targets
by TRUE, and FALSE for not-on-targets.}
}
\value{
A list of lists of data frames with the original columns and an
additional \code{lfdr} column. The \code{is_target} column is optionally appended
only when \code{known_target_df} is not NULL.
}
\description{
Append local false discovery (lfdr) score to the statistical testing results.
In addition, append is_target column if known on-target data is available.
}
