% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_t_test.R
\name{compute_mslip_on_stdreport}
\alias{compute_mslip_on_stdreport}
\title{Run MSstatsLiP comparing two conditions on a standard report}
\usage{
compute_mslip_on_stdreport(report)
}
\arguments{
\item{report}{Fragment ion report with the columns experiment, condition,
replicate, protein_id, precursor_id, fragment_id, and fragment_peak_area.
The condition column is supposed to consist of two conditions.}
}
\value{
A data frame of the test results. Columns are
experiment, protein_id, peptide_sequence, statistic, df, p.value, estimate,
Label, and miscellaneous columns from MSstatsLiP::groupComparisonLiP().
And rows are precursors. Note that precursor_id column is missing.
}
\description{
Run MSstatsLiP comparing two conditions on a standard report
}
