% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_t_test.R
\name{compute_paired_on_group}
\alias{compute_paired_on_group}
\title{Wrapper function for paired t-test}
\usage{
compute_paired_on_group(
  groupdf,
  column_x = "Log10NormalizedPeakArea.x",
  column_y = "Log10NormalizedPeakArea.y"
)
}
\arguments{
\item{groupdf}{A subset of fragment ion report which contains two columns for
log10-transformed fragment ion quantities of conditions x and y}

\item{column_x}{column name for x}

\item{column}{y column name for y}
}
\value{
a data frame of the paired t-test results
}
\description{
Wrapper function for paired t-test
}
