% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_t_test.R
\name{compute_rots_on_stdreport}
\alias{compute_rots_on_stdreport}
\title{Run ROTS comparing two conditions on a standard report}
\usage{
compute_rots_on_stdreport(report)
}
\arguments{
\item{report}{fragment ion report with the columns experiment, condition,
replicate, protein_id, precursor_id, fragment_id, and fragment_peak_area.
The condition column is supposed to consist of two conditions.}
}
\value{
a data frame of the test results. Columns are
experiment, protein_id, precursor_id, statistic, df, p.value, estimate.
And rows are precursors.
}
\description{
Run ROTS comparing two conditions on a standard report
}
