% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{compute_sensitivity_result}
\alias{compute_sensitivity_result}
\title{Compute sensitivity and specificity from statistical significance result}
\usage{
compute_sensitivity_result(
  lfdr_result,
  is_target_column = "is_target",
  q_value_column = "lfdr",
  group_column = c("experiment", "protein_id"),
  random_seed = 11111
)
}
\arguments{
\item{lfdr_result}{A list of lists of statistical test result tables,
which are obtained by run_ttest(), but each result table has two
additional columns specified by 'is_target_column' and 'q_value_column'.}

\item{is_target_column}{Name of the column having a boolean vector for
labeling true target.}

\item{q_value_column}{Name of the column having significance levels: A
smaller value implies more significance.}

\item{group_column}{Names of the columns for grouping the rows. Defaults
to c('experiment', 'protein_id')}

\item{random_seed}{Random seed. Used for shuffling input values.}
}
\value{
A list of lists of data frames having nine metrics:
total, total_pos, total_neg, count_p, count_tp, count_fp, count_fn,
specificity, sensitivity. The count_p is a vector of the number of
positives, which is set to increase from 1 to total.
}
\description{
Compute sensitivity and specificity from statistical significance result
}
