% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_mq_to_standard}
\alias{convert_mq_to_standard}
\title{Convert MaxQuant report into standard report}
\usage{
convert_mq_to_standard(evidence, msms, annotation = NULL)
}
\arguments{
\item{evidence}{MaxQuant evidence.txt data}

\item{msms}{MqxQuant msms.txt data}

\item{annotation}{Data frame with columns Raw.file, Condition, BioReplicate,
IsotopeLabelType}
}
\value{
A standard report with columns condition, replicate, experiment,
protein_id, precursor_id, fragment_id, fragment_peak_area
}
\description{
Convert MaxQuant report into standard report. MaxQuant report consists of
evidence.txt data, msms.txt data, and annotation data.
}
