% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_standard_to_ms}
\alias{convert_standard_to_ms}
\title{Convert standard report into MSstats format}
\usage{
convert_standard_to_ms(report)
}
\arguments{
\item{report}{A standard report. Required columns are condition, replicate,
experiment, protein_id, precursor_id, fragment_id, fragment_peak_area.}
}
\value{
A standard report with columns ProteinName, PeptideSequence,
PrecursorCharge, FragmentIon, ProductCharge, Condition, BioReplicate, Run,
Intensity
}
\description{
No preprocessing is carried out, but column names are changed.
}
