% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_standard_to_mslip}
\alias{convert_standard_to_mslip}
\title{Convert standard report into MSstatsLiP format}
\usage{
convert_standard_to_mslip(report, drop_experiment = FALSE)
}
\arguments{
\item{report}{A standard report. Required columns are condition, replicate,
experiment, protein_id, precursor_id, fragment_id, fragment_peak_area.}

\item{drop_experiment}{If TRUE, \code{experiment} column is removed from output.
Default is FALSE.}
}
\value{
An MSstatsLiP format with columns ProteinName, PeptideSequence,
PrecursorCharge, FragmentIon, ProductCharge, Condition, BioReplicate, Run,
Intensity, FULL_PEPTIDE
}
\description{
Column names are changed, and input report is preprocessed in the same way as
MSstatsLiP converter.
}
