% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_t_test.R
\name{cov_diff_shrink}
\alias{cov_diff_shrink}
\title{Covariance of diff between each pair of fragment ions based on shrinkage.}
\usage{
cov_diff_shrink(
  dat_con1,
  dat_con2,
  lambda_var_con1,
  lambda_var_con2,
  cov_equal = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dat_con1, dat_con2}{data matrices of normalized fragment ion peak area,
each row is a replicate, each column is a fragment ion}

\item{lambda_var_con1, lambda_var_con2}{the variance shrinkage intensities
for con1 and con2, respectively. If not specified (default), they are
estimated by corpcor::cov.shrink function.}

\item{cov_equal}{TRUE if equal covariance is assumed and a pooled
covariance matrix is computed.}

\item{verbose}{TRUE to print messages. Default is FALSE.}
}
\value{
covariance matrix of mean difference
}
\description{
A diff for an ion i is defined by d_{i} = xbar_con1{i} - xbar_con2{i}, where
xbar_con1 and xbar_con2 are the sample means for condition 1 and 2,
respectively. A covariance of diff between fragment ion i and j is
cov(d_{i}, d_{j})
= cov(xbar_con1{i}, xbar_con1{j}) + cov(xbar_con2{i}, xbar_con2{j})
}
\examples{
  dat_x = matrix(rnorm(12), 4, 3)
  dat_y = matrix(rnorm(12), 4, 3)
  cov_diff_shrink(dat_x, dat_y)
}
