% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_annotation_df}
\alias{create_annotation_df}
\title{Create annotation data for MSstats}
\usage{
create_annotation_df(
  evidence,
  n_replicates_per_condition = NULL,
  file_column = "Raw.file",
  exp_column = "Experiment"
)
}
\arguments{
\item{evidence}{A data frame including the columns named by \code{file_column} and
\code{exp_column}.}

\item{n_replicates_per_condition}{Number of replicates per condition.
If not NULL, it is assumed that Experiment is n_condition x
n_replicates_per_condition.}

\item{file_column}{Column name for run files}

\item{exp_column}{Column name for Experiment}
}
\value{
Annotation data frame.
}
\description{
Annotation is required by MSstats for importing reports, such as
MaxQuant. Annotation data consists of Condition, Raw.file, BioReplicate,
and IsotopeLabelType. Condition column is inferred from Experiment column.
}
