% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{create_mq_annotation_file}
\alias{create_mq_annotation_file}
\title{Create MaxQuant annotation file for MSstats}
\usage{
create_mq_annotation_file(
  mq_evidence,
  annotation_path = "data/sample_annotation.txt",
  n_replicates_per_condition = NULL
)
}
\arguments{
\item{mq_evidence}{A data frame for a MaxQuant evidence.txt data}

\item{annotation_path}{Path to the annotation txt file to write. If a file
exists at the designated path, user will be asked to make sure to write
on that path.}

\item{n_replicates_per_condition}{Number of replicates per condition.
Defaults to NULL. If NULL, Condition is inferred from Experiment.}
}
\value{
Path to the created annotation.txt file. Empty string "" if failed.
}
\description{
Create an annotation file in tsv format. Condition column is inferred from
the Experiment column of MaxQuant evidence data.
}
