% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_report.R
\docType{data}
\name{default_params}
\alias{default_params}
\title{Define default parameters for sampling distributions}
\format{
An object of class \code{list} of length 18.
}
\usage{
default_params
}
\description{
Define default parameters for sampling distributions
}
\keyword{datasets}
