% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mixture.R
\name{dexpbeta10}
\alias{dexpbeta10}
\title{Exponential-beta density function with base 10}
\usage{
dexpbeta10(x, shape1 = 1, shape2 = 1)
}
\arguments{
\item{x}{non-positive values}

\item{shape1, shape2}{shape parameters}
}
\value{
density function values
}
\description{
Exponential-beta density function with base 10
}
