% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.comp_cov_uneq_repl}
\alias{.comp_cov_uneq_repl}
\title{Compute cov_unequal_replicates}
\usage{
.comp_cov_uneq_repl(x)
}
\arguments{
\item{x}{a vector of mean precursor quantities in log10-scale}
}
\value{
the cov_unequal_replicates value
}
\description{
Compute cov_unequal_replicates from data values
}
