% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{.compute_sensitivity_table}
\alias{.compute_sensitivity_table}
\title{Compute sensitivity and specificity from statistical significance result}
\usage{
.compute_sensitivity_table(is_target, q_values, random_seed = 1111)
}
\arguments{
\item{is_target}{A boolean vector for labeling true target.}

\item{q_values}{Significance levels. Smaller value implies more significance.}

\item{random_seed}{Random seed. Used for shuffling input values.}
}
\value{
data frame of nine metrics: total, total_pos, total_neg, count_p,
count_tp, count_fp, count_fn, specificity, sensitivity. The count_p is
a vector of the number of positives, which increases from 1 to total.
}
\description{
Compute sensitivity and specificity from statistical significance result
}
