% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mixture.R
\name{fit_mixture_normal_expbeta}
\alias{fit_mixture_normal_expbeta}
\title{Fit a mixture of normal and expbeta distributions by EM algorithm}
\usage{
fit_mixture_normal_expbeta(
  x,
  init_params,
  max_iter = 30,
  index_fixed_params = c(1),
  xlim_max = NULL
)
}
\arguments{
\item{x}{a vector of data values}

\item{init_params}{a vector of initial parameter values}

\item{index_fixed_params}{index for fixed parameters. A subset of {1,..., 4}}
}
\value{
a vector of fitted parameters and component weights
}
\description{
Fit a mixture of normal and expbeta distributions by EM algorithm
}
