% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_t_test.R
\name{independent_t_test}
\alias{independent_t_test}
\title{Independent (two) sample t-test for modified peptide using aggregated
quantity}
\usage{
independent_t_test(
  quantity1,
  quantity2,
  var_equal = TRUE,
  conf_level = 0.95,
  verbose = FALSE
)
}
\arguments{
\item{quantity1, quantity2}{numeric vectors of normalized peptide quantity}

\item{var_equal}{TRUE to assume equal variance between conditions. Default
is TRUE.}

\item{conf_level}{Confidence level. Default is 0.95.}

\item{verbose}{TRUE to print messages. Default is FALSE.}
}
\value{
list of statistic, df, p.value, and estimate.
}
\description{
Independent (two) sample t-test for modified peptide using aggregated
quantity
}
\examples{
independent_t_test(rnorm(4), rnorm(4))
}
