% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paired_t_test.R
\name{paired_t_test}
\alias{paired_t_test}
\title{Paired t-test for modified peptide using fragment ion peak area}
\usage{
paired_t_test(quantity1, quantity2, conf_level = 0.95, verbose = FALSE)
}
\arguments{
\item{quantity1, quantity2}{numeric vectors of normalized fragment peak area}

\item{conf_level}{Confidence level. Default is 0.95.}

\item{verbose}{TRUE to print messages. Default is FALSE.}
}
\value{
list of statistic, df, p.value, and estimate.
}
\description{
Paired t-test for modified peptide using fragment ion peak area
}
\examples{
paired_t_test(rnorm(10), rnorm(10))
}
