% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbetamixture.R
\name{rbetamixture}
\alias{rbetamixture}
\title{Generate random samples from a mixture of Beta distributions}
\usage{
rbetamixture(n, shape1s = 1, shape2s = 1)
}
\arguments{
\item{n}{number of observations}

\item{shape1s, shape2s}{vectors of non-negative parameters of component
beta distributions. Length of each vector is the number of component beta
distributions.}
}
\value{
a length-n vector of generated values
}
\description{
Generate random samples from a mixture of Beta distributions
}
