% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_ttests}
\alias{run_ttests}
\title{Run t-test methods}
\usage{
run_ttests(
  report,
  method_names = NULL,
  boot_denom_eps = 0.3,
  base_condition = NULL
)
}
\arguments{
\item{report}{Fragment ion report with the columns experiment, condition,
replicate, protein_id, precursor_id, fragment_id, fragment_peak_area.}

\item{method_names}{A vector of characters for the method names to execute.
If NULL, it is set to all the method names available. Defaults to NULL.}

\item{boot_denom_eps}{A parameter for shrinkage t-test.}

\item{base_condition}{Base condition. If NULL, the first condition among the
unique list of the condition column is selected. Default to NULL.}
}
\value{
A list of analysis results of the given methods. For each method,
the result is a list of the pairwise comparisons.
}
\description{
Run t-test methods for the pairwise comparisons:
cond1 / cond2, cond1 / cond3, ..., cond1 / cond_C,
where cond1 is called the base condition.
Each pairwise comparisons are performed on each
experiment x protein_id x precursor_id.
}
