% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_t_test.R
\name{se_diff_shrink}
\alias{se_diff_shrink}
\title{Standard error of estimate for shrinkage-based t-test statistic.}
\usage{
se_diff_shrink(
  dat_con1,
  dat_con2,
  lambda_var,
  cov_unequal_replicates = 0,
  cov_equal = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{dat_con1, dat_con2}{Data matrices of normalized fragment ion peak area
in logarithmic scale, \eqn{x_{1ir}^p, x_{2ir}^p}. Each row is a replicate,
each column is a fragment ion.}

\item{lambda_var}{Variance shrinkage intensities. If not specified
(default), they are estimated by corpcor::cov.shrink function.}

\item{cov_unequal_replicates}{covariance of input variables whose replicates
are not equal, \eqn{Cov(x_{ci_1r_1}^p, x_{ci_2r_2}^p), r_1 \neq r_2}.}

\item{cov_equal}{TRUE if equal covariance is assumed and a pooled
covariance matrix is computed.}

\item{verbose}{TRUE to print messages. Default is FALSE.}
}
\value{
standard error of the sum of mean differences
}
\description{
Standard error appears in the the denominator of the test statistic formula.
In the shrinkage-based t-test, the estimate of group mean difference is the
sum of \eqn{d_i} across fragment ions \eqn{i}, where
\eqn{d_i = \bar{x}_{1i} - \bar{x}_{2i}},
with \eqn{\bar{x}_{1i}} and \eqn{\bar{x}_{2i}} representing the mean
quantities for fragment ion \eqn{i} in groups 1 and 2, respectively.
}
