% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_t_test.R
\name{shrinkage_t_test_statistic}
\alias{shrinkage_t_test_statistic}
\title{Shrinkage-based t-test statistic}
\usage{
shrinkage_t_test_statistic(
  dat_con1,
  dat_con2,
  cov_equal,
  denom_eps,
  lambda_var
)
}
\arguments{
\item{dat_con1, dat_con2}{Data matrices of normalized fragment ion peak area
in logarithmic scale. Each row is a replicate, and each column is a
fragment ion.}

\item{cov_equal}{TRUE if equal covariance is assumed and a pooled
covariance matrix is computed.}

\item{denom_eps}{A small constant to be added to the denominator (i.e.,
standard error of the numerator) of the statistic. This constant is useful
for bootstrapping with a small sample size.}

\item{lambda_var}{Variance shrinkage intensity. If not specified (default),
they are estimated by corpcor::cov.shrink function.}
}
\value{
shrinkage-based t-test statistic
}
\description{
Shrinkage-based t-test statistic
}
