% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_report.R
\name{simulate_fragment_ion_report}
\alias{simulate_fragment_ion_report}
\title{Generate a simulated fragment ion report based on a hierarchical graphical
model}
\usage{
simulate_fragment_ion_report(params, seed = 100)
}
\arguments{
\item{params}{A list of parameters in the same format as default_params}

\item{seed}{An integer for random seed}
}
\value{
A data frame of a fragment ion report
}
\description{
A fragment ion report consists of the following columns:
\itemize{
\item experiment, condition, replicate, protein_id, precursor_id,
precursor_quantity, fragment_id, fragment_peak_area
}
}
\details{
The precursor_quantity and fragment_peak_area columns are raw values, i.e.,
not log-transformed.
}
\examples{
report <- simulate_fragment_ion_report(default_params)
x1 <- report[
    report$condition == "CON1" & report$fragment_id == "FRAG1"
]$precursor_quantity
par(mfrow = c(1,2))
hist(log10(x1), main = "Histogram", xlab = "Precursor Quantity, log10",
     xlim = c(2, 6))
qqnorm(log10(x1))
qqline(log10(x1))
}
