% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{sw_normalize_values_df}
\alias{sw_normalize_values_df}
\title{Normalize values to standard normal distribution by sliding window}
\usage{
sw_normalize_values_df(
  values_df,
  value_column = "precursor_quantity",
  sliding_column = "score",
  window_size = 1e+05,
  step_size = 50000,
  reconstruct = TRUE,
  use_logvalues = TRUE,
  normalized_value_column = "normalized_precursor_quantity",
  normalization_factor_column = "normalization_factor"
)
}
\arguments{
\item{values_df}{data frame with value_column and sliding_column}

\item{value_column}{name of the value column}

\item{sliding_column}{name of the score column that sliding window runs over}

\item{window_size}{window size for sliding column with which an empirical
cdf is estimated}

\item{step_size}{step size for sliding column for which corresponding
value_column is normalized}

\item{reconstruct}{If TRUE, the normalized_value_column takes the robust mean
and robust standard deviation of the original value_column. If FALSE, the
mean and standard deviation are set to 0 and 1.}

\item{use_logvalues}{If TRUE, the value column is log10-transformed,
normalized, and then transformed back to the original exponential scale.}

\item{normalized_value_column}{name of the normalized value column}

\item{normalization_factor_column}{name of the normalization factor column}
}
\value{
data frame with normalized values set in the normalized_value_column
}
\description{
Normalize values to standard normal distribution by sliding window
}
