% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_contingency_tables}
\alias{compute_contingency_tables}
\title{Compute contingency tables}
\usage{
compute_contingency_tables(results_run_ttests, alpha = 0.05)
}
\arguments{
\item{results_run_ttests}{results of run_ttest function, which is a named
list of analysis results of each method.}

\item{alpha}{significance level}
}
\value{
list of contingency tables for every comparisons. Each element of
the list is a table, whose columns are the t-test methods and the rows are
TRUE/FALSE representing whether H0 is rejected or not.
}
\description{
Compute contingency tables based on the results of running t-test methods.
}
