% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_diff.R
\name{cov_diff}
\alias{cov_diff}
\title{Covariance of diff between each pair of fragment ions.}
\usage{
cov_diff(dat_con1, dat_con2, verbose = FALSE)
}
\arguments{
\item{dat_con1, dat_con2}{data matrices of normalized fragment ion peak area,
each row is a replicate, each column is a fragment ion}

\item{verbose}{TRUE to print messages. Default is FALSE.}
}
\value{
covariance matrix of mean difference
}
\description{
A diff for an ion i is defined by d_{i} = xbar_con1{i} - xbar_con2{i}, where
xbar_con1 and xbar_con2 are the sample means for condition 1 and 2,
respectively. A covariance of diff between fragment ion i and j is
cov(d_{i}, d_{j})
= cov(xbar_con1{i}, xbar_con1{j}) + cov(xbar_con2{i}, xbar_con2{j})
}
\examples{
  dat_x = matrix(rnorm(12), 4, 3)
  dat_y = matrix(rnorm(12), 4, 3)
  cov_diff(dat_x, dat_y)
}
