% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_t_test.R
\name{cov_diff_shrink}
\alias{cov_diff_shrink}
\title{Covariance of diff between each pair of fragment ions based on shrinkage.}
\usage{
cov_diff_shrink(
  dat_con1,
  dat_con2,
  lambda_var_con1,
  lambda_var_con2,
  cov_equal = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dat_con1, dat_con2}{data matrices of normalized fragment ion peak area
in logarithmic scale. Each row is a replicate, each column is a fragment
ion.}

\item{lambda_var_con1, lambda_var_con2}{the variance shrinkage intensities
for con1 and con2, respectively. If not specified (default), they are
estimated by corpcor::cov.shrink function.}

\item{cov_equal}{TRUE if equal covariance is assumed and a pooled
covariance matrix is computed.}

\item{verbose}{TRUE to print messages. Default is FALSE.}
}
\value{
covariance matrix of mean difference
}
\description{
A diff for an ion \eqn{i} is defined by
\eqn{d_i = \bar{x}_{1i} - \bar{x}_{2i}}, where
\eqn{\bar{x}_{1i}} and \eqn{\bar{x}_{1i}} are the sample means for
conditions 1 and 2, respectively. A covariance of diff between fragment ions
\eqn{i} and \eqn{j} is
\deqn{cov(d_i, d_j)
   = cov(\bar{x}_{1i}, \bar{x}_{1j}) +
   cov(\bar{x}_{2i}, \bar{x}_{2j})}
}
\examples{
  dat_x <- cbind(1 : 4, c(1, 3, 2, 4), 4 : 1)
  dat_y <- cbind(4 : 1, c(4, 2, 3, 1), 1 : 4)
  cov_diff_shrink(dat_x, dat_y)
}
