% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{line_plot_contingency_tables}
\alias{line_plot_contingency_tables}
\title{Line plot for contingency tables}
\usage{
line_plot_contingency_tables(
  x,
  tables,
  rejected = TRUE,
  scale_factor = 1,
  add_legend = FALSE,
  legend_coord = "topright",
  legend_cex = 1,
  ...
)
}
\arguments{
\item{x}{horizontal coordinates of the line plots}

\item{tables}{list contingency tables. Lengths of x and tables must be equal.}

\item{rejected}{TRUE if number of rejected hypothesis is plotted in y axis.}

\item{scale_factor}{a numeric factor that is multiplied to data values.}

\item{add_legend}{If not FALSE, legend is added in the plot}

\item{legend_coord}{coordinate of the legend}

\item{legend_cex}{cex for the legend}

\item{...}{Additional arguments to graphics::matplot}
}
\description{
Line plot for contingency tables
}
