% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{local_fdr}
\alias{local_fdr}
\title{Compute local false discovery rates from p-values}
\usage{
local_fdr(pvalues, known_target_df, eps = 5e-16)
}
\arguments{
\item{pvalues}{A vector of p-values}
}
\value{
A vector of local fdr values, of the same length as pvalues
}
\description{
The p-values are transformed to z-values, then, lfdr values are computed.
}
