% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_ttests}
\alias{run_ttests}
\title{Run t-test methods}
\usage{
run_ttests(report, boot_denom_eps = 0.5, base_condition = NULL)
}
\arguments{
\item{report}{fragment ion report with the columns experiment, condition,
replicate, protein_id, precursor_id, fragment_id, fragment_peak_area.}

\item{boot_denom_eps}{a parameter for shrinkage t-test}
}
\value{
a list of analysis results of three t-test methods, where the result
of each method is a list for the pairwise comparisons.
}
\description{
Run t-test methods for the pairwise comparisons: 1-2, 1-3, ..., 1-C.
Each pairwise comparisons are performed on each
experiment x protein_id x precursor_id.
}
