% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_t_test.R
\name{se_diff_shrink}
\alias{se_diff_shrink}
\title{Standard error of estimate for shrinkage-based t-test statistic.}
\usage{
se_diff_shrink(
  dat_con1,
  dat_con2,
  lambda_var,
  cov_equal = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dat_con1, dat_con2}{Data matrices of normalized fragment ion peak area
in logarithmic scale. Each row is a replicate, each column is a fragment
ion.}

\item{lambda_var}{Variance shrinkage intensities. If not specified
(default), they are estimated by corpcor::cov.shrink function.}

\item{cov_equal}{TRUE if equal covariance is assumed and a pooled
covariance matrix is computed.}

\item{verbose}{TRUE to print messages. Default is FALSE.}
}
\value{
standard error of the sum of mean differences
}
\description{
The estimate is the sum of diff over fragment ions, where diff for a fragment
ion is defined by \eqn{d_i = \bar{x}_{1i} - \bar{x}_{2i}}, and it
constitutes the numerator of the test statistic formula.
}
