% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkage_t_test.R
\name{shrinkage_t_test}
\alias{shrinkage_t_test}
\title{Shrinkage-based t-test for testing group mean differeneces in peptide mean
quantity using fragment ion peak area}
\usage{
shrinkage_t_test(
  dat_con1,
  dat_con2,
  num_boot = 100,
  cov_equal = TRUE,
  boot_denom_eps = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{dat_con1, dat_con2}{data matrices of normalized fragment ion peak area,
each row is a replicate, each column is a fragment ion.}

\item{num_boot}{number of bootstrap replicates. Default is 200.}

\item{cov_equal}{TRUE if equal covariance is assumed and a pooled
covariance matrix is computed. Default is TRUE.}

\item{boot_denom_eps}{a small constant to be added to the denominator (i.e.,
standard error of the numerator) of the bootstrapped statistic. This
constant is useful for bootstrapping with a small sample size.}

\item{verbose}{TRUE to print messages. Default is FALSE.}
}
\value{
list of statistic, df, p.value, and estimate.
}
\description{
Shrinkage-based t-test for testing group mean differeneces in peptide mean
quantity using fragment ion peak area
}
\examples{
  dat_x = matrix(rnorm(12), 4, 3)
  dat_y = matrix(rnorm(12), 4, 3)
  shrinkage_t_test(dat_x, dat_y)
}
